Refactor GameReducer into modular structure
- Split monolithic GameReducer.ts (857 lines) into category-specific reducers
- Implemented modular pattern similar to ResourceSystem for better maintainability
- Created src/core/reducers/ directory structure with focused reducer modules
- ResourceReducer split into: computing, funding, influence, and data sub-modules
- Extracted smaller reducers as single files: meta, deployment, event, world, competitor, settings
- Updated all imports to use new modular structure
- Fixed type issues and aligned research reducer with actual interfaces
- Added documentation for reducer organization
- Updated ROADMAP to reflect completed refactoring

This refactoring improves code organization, maintainability, and follows established patterns.
The research reducer remains in GameReducer.ts temporarily, pending Phase 3 modularization.
