+++ ROADMAP.md
@@ -139,7 +139,7 @@ The roadmap will be updated as development progresses to reflect new priorities,
 **Rationale:** Prevent technical debt before building more complex systems
 - [x] Refactor ResearchTreeView.ts (1064 lines) into smaller components
 - [x] Split ResourceSystem.ts (802 lines) into smaller modules
-- [ ] Break down GameReducer.ts (850 lines) into category-specific reducers
+- [x] Break down GameReducer.ts (850 lines) into category-specific reducers
 - [x] Clarify UIComponent architecture and remove unused code
 - [ ] Audit and clean up CSS redundancy
