+++ src/core/reducers/index.ts
@@ -1,0 +1,14 @@
+/**
+ * Central export point for all reducers
+ */
+
+export { gameReducer } from './gameReducer';
+export { metaReducer } from './metaReducer';
+export { resourceReducer } from './resourceReducer';
+export { deploymentReducer } from './deploymentReducer';
+export { eventReducer } from './eventReducer';
+export { worldReducer } from './worldReducer';
+export { competitorReducer } from './competitorReducer';
+export { settingsReducer } from './settingsReducer';
+
+// TODO: Export researchReducer once it's modularized
